Validate that ProjectUrl points to an existing web site.
If not, make sure we don't link to the web site in the audit report. This step take a little over 2 minutes which means that it will take a longer to generate the reports but it's not a big deal since the audit is executed unattended twice a day. However, it is a big deal when executing the audit locally for testing/debugging purposes. Therefore this step will be skipped when "-excludeslowsteps" is specified on the command line.

Resolves #102