Make sure we use same branch when comparing fork with upstream.
When a fork is newly created, the 'DefaultBranch' property is not set to the expected value and defaults to 'master'. This is a problem if the default branch is anything other than master (such as 'develop' for example) and you are comparing the fork with its parent because you end up attempting to compare the parent's 'develop' with the fork's 'master'.