Handle cases where assembly name is unknown when generating YAML
In other words, use `*.dll` rather than an empty string if the addin discoverer was unable to determine the name of the assembly. This happens when there are multiple assemblies in a nuget package, these assemblies do not follow the naming convention and none of the have methods decorated with either CakeMethodAliasAttribute or CakePropertyAliasAttribute. For example: modules.