Restore the GetGithubMetadata step to confirm it's the culprit of timeout issue
The timeout issue has stopped following the exclusion of "slow" steps (i.e.: GetGithubMetadata and CheckUsingCakeRecipe). This means that one of the 4 AnalysisResult properties set by CheckUsingCakeRecipe or one of the three set by GetGithubMetadata is the culprit of the timeout issue. Out of the 7 properties set by these two steps, 6 are simple 'bool' or 'int' or 'string'. The seventh property is a Dictionary<string, Stream>' called 'RepoContent' which I suspect is the root of this issue and should be excluded from serialization.